+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 - [#187](https://github.com/MiSchroe/klf-200-api/issues/187) Fix potential memory leaks.
 - **BREAKING CHANGE:** Changed internal Disposable interface to use [Explicit Resource Management](https://github.com/tc39/proposal-explicit-resource-management).
+- Add [debug](https://github.com/debug-js/debug#readme) instrumentation to the code.
 
 ## 6.0.0 (2024-12-30)